CruiseComputer: calculate three point distance
The cruise computer, calculating cruise distance, previously only used distance from start to current location,
so if the glider glides into a turn point and out again, the gradient is meaningless.

This new code uses a convex hull with quick search for an intermediate point to maximise the distance
between start to intermediate to current location.